backend/backend/requirements.txt
# ============================================================
# PQRSD Inteligente - Alcaldía de Medellín
# Backend Dependencies
# ============================================================

# --- Web Framework ---
flask==3.0.3
flask-cors==4.0.1

# --- Environment Variables ---
python-dotenv==1.0.1

# --- Supabase (base de datos) ---
supabase==2.5.3

# --- Groq IA ---
groq>=0.9.0

# --- Utilidades de peticiones HTTP ---
httpx==0.27.0
requests==2.32.3
